+++ redis/src/test/scala/zio/redis/internal/RespValueSpec.scala
@@ -2,7 +2,6 @@ package zio.redis.internal
 
 import zio.Chunk
 import zio.redis._
-import zio.redis.internal.RespValue
 import zio.test.Assertion._
 import zio.test._
